refactor(rust): Split encoding.rs into modular structure
