+++ lib/src/encoding/mod.rs
@@ -1,0 +1,9 @@
+//! String to byte array encoding functions
+
+mod bytes_to_string;
+mod helpers;
+mod string_to_bytes;
+
+// Re-export public FFI functions
+pub use bytes_to_string::{bytes_to_string, bytes_to_string_lenient};
+pub use string_to_bytes::string_to_bytes;
